requirements.txt
# PROTONORD Shopify Management System - Python Dependencies

# Core dependencies for Shopify API and data processing
requests==2.32.5
psycopg2-binary==2.9.10
pandas==2.3.2
pytz==2025.2

# Data visualization and reporting
plotly==6.3.0

# Date and time handling
python-dateutil>=2.8.0

# JSON processing and data validation
jsonschema>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Logging and monitoring
coloredlogs>=15.0

# File and path utilities
pathlib>=1.0.0

# HTTP server for dashboard
# Note: Python's built-in http.server is used, no additional package needed

# Database utilities
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration management
configparser>=5.0.0

# Security and encryption
cryptography>=41.0.0

# System monitoring
psutil>=5.9.0

# Testing (optional, for development)
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality (optional, for development)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development utilities (optional)
ipython>=8.0.0
jupyter>=1.0.0